Move getTodos api to houses folders
