+++ server/src/app.module.ts
@@ -3,12 +3,14 @@ import { ConfigModule } from '@nestjs/config';
 import { UserModule } from './user/user.module';
 import { AuthModule } from './auth/auth.module';
 import { PrismaModule } from './prisma/prisma.module';
+import { HouseModule } from './house/house.module';
 
 @Module({
   imports: [
     ConfigModule.forRoot({ isGlobal: true }),
     UserModule,
     AuthModule,
+    HouseModule,
     PrismaModule,
   ],
 })
