+++ server/src/house/house.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { HouseController } from './house.controller';
+import { HouseService } from './house.service';
+
+@Module({
+  controllers: [HouseController],
+  providers: [HouseService],
+})
+export class HouseModule {}
