+++ server/src/user/user.controller.ts
@@ -1,10 +1,4 @@
-import {
-  Controller,
-  Get,
-  Param,
-  ParseIntPipe,
-  UseGuards,
-} from '@nestjs/common';
+import { Controller, Get, UseGuards } from '@nestjs/common';
 import { UserService } from './user.service';
 import { AuthGuard } from '../auth/guard';
 import { GetUser } from '../auth/decorator';
@@ -17,14 +11,5 @@ export class UserController {
   @Get()
   getMe(@GetUser('id') userId: number) {
     return this.userService.getMe(userId);
-  }
-
-  @UseGuards(AuthGuard)
-  @Get(':houseId/todos')
-  getTodos(
-    @GetUser('id') userId: number,
-    @Param('houseId', ParseIntPipe) houseId: number,
-  ) {
-    return this.userService.getTodos(userId, houseId);
   }
 }
